break user_relation cycle
