+++ front/lib/api/assistant/configuration.ts
@@ -51,7 +51,6 @@ import {
   isGlobalAgentId,
 } from "@app/lib/api/assistant/global_agents";
 import { agentConfigurationWasUpdatedBy } from "@app/lib/api/assistant/recent_authors";
-import { agentUserListStatus } from "@app/lib/api/assistant/user_relation";
 import { compareAgentsForSort } from "@app/lib/assistant";
 import type { Authenticator } from "@app/lib/auth";
 import { getPublicUploadBucket } from "@app/lib/file_storage";
@@ -364,6 +363,29 @@ async function fetchAgentConfigurationsForView(
       }
       assertNever(agentsGetView);
   }
+}
+
+export function agentUserListStatus({
+  agentConfiguration,
+  listStatusOverride,
+}: {
+  agentConfiguration: LightAgentConfigurationType;
+  listStatusOverride: AgentUserListStatus | null;
+}): AgentUserListStatus {
+  if (listStatusOverride === null) {
+    switch (agentConfiguration.scope) {
+      case "global":
+      case "workspace":
+      case "private":
+        return "in-list";
+      case "published":
+        return "not-in-list";
+      default:
+        assertNever(agentConfiguration.scope);
+    }
+  }
+
+  return listStatusOverride;
 }
 
 async function fetchWorkspaceAgentConfigurationsForView(
